commit with all changes in yaaic

+++ app/src/main/java/org/yaaic/menu/ServerPopupMenu.java
@@ -55,20 +55,42 @@ public class ServerPopupMenu extends PopupMenu implements PopupMenu.OnMenuItemCl
     public boolean onMenuItemClick(MenuItem item) {
         int id = item.getItemId();
 
-        switch (id) {
-            case R.id.edit:
-                listener.onEditServer(server);
-                break;
-            case R.id.delete:
+        if(id == R.id.edit) {
+            listener.onEditServer(server);
+        }else{
+            if(id == R.id.delete) {
                 listener.onDeleteServer(server);
-                break;
-            case R.id.connect:
-                listener.onConnectToServer(server);
-                break;
-            case R.id.disconnect:
-                listener.onDisconnectFromServer(server);
-                break;
+            }else{
+                if(id == R.id.connect) {
+                    listener.onConnectToServer(server);
+                }else{
+                    if(id == R.id.disconnect) {
+                        listener.onDisconnectFromServer(server);
+                    }else{
+
+                    }
+                }
+            }
         }
+
+
+
+
+
+//        switch (id) {
+//            case R.id.edit:
+//                listener.onEditServer(server);
+//                break;
+//            case R.id.delete:
+//                listener.onDeleteServer(server);
+//                break;
+//            case R.id.connect:
+//                listener.onConnectToServer(server);
+//                break;
+//            case R.id.disconnect:
+//                listener.onDisconnectFromServer(server);
+//                break;
+//        }
 
         return true;
     }
